rename discovery.k8s to discovery.kubernetes. Move much of implementation to core discovery module
