+++ component/all/all.go
@@ -2,7 +2,7 @@
 package all
 
 import (
-	_ "github.com/grafana/agent/component/discovery/kubernetes"                  // Import discovery.k8s
+	_ "github.com/grafana/agent/component/discovery/kubernetes"                  // Import discovery.kubernetes
 	_ "github.com/grafana/agent/component/discovery/relabel"                     // Import discovery.relabel
 	_ "github.com/grafana/agent/component/local/file"                            // Import local.file
 	_ "github.com/grafana/agent/component/prometheus/integrations/node_exporter" // Import prometheus.integration.node_exporter
