+++ component/discovery/kubernetes/k8s.go
@@ -1,9 +1,6 @@
 package kubernetes
 
 import (
-	"context"
-	"sync"
-
 	"github.com/grafana/agent/component"
 	"github.com/grafana/agent/component/common/config"
 	"github.com/grafana/agent/component/discovery"
@@ -12,14 +9,19 @@ import (
 
 func init() {
 	component.Register(component.Registration{
-		Name:    "discovery.k8s",
+		Name:    "discovery.kubernetes",
 		Args:    SDConfig{},
-		Exports: Exports{},
+		Exports: discovery.Exports{},
 
 		Build: func(opts component.Options, args component.Arguments) (component.Component, error) {
-			return New(opts, args.(SDConfig))
+			return discovery.New(opts, args, newK8s)
 		},
 	})
+}
+
+var newK8s discovery.DiscoveryCreator = func(args component.Arguments, opts component.Options) (discovery.Discoverer, error) {
+	newArgs := args.(SDConfig)
+	return promk8s.New(opts.Logger, newArgs.Convert())
 }
 
 // SDConfig is a conversion of discover/kubernetes/SDConfig to be compatible with flow
@@ -86,79 +88,4 @@ func (sc *SelectorConfig) convert() *promk8s.SelectorConfig {
 		Label: sc.Label,
 		Field: sc.Field,
 	}
-}
-
-// Exports holds values which are exported by the discovery.k8s component.
-type Exports struct {
-	Targets []discovery.Target `river:"targets,attr"`
-}
-
-// Component implements the discovery.k8s component.
-type Component struct {
-	opts component.Options
-
-	discMut       sync.Mutex
-	latestDisc    discovery.Discoverer
-	newDiscoverer chan struct{}
-}
-
-// New creates a new discovery.k8s component.
-func New(o component.Options, args SDConfig) (*Component, error) {
-	c := &Component{
-		opts: o,
-		// buffered to avoid deadlock from the first immediate update
-		newDiscoverer: make(chan struct{}, 1),
-	}
-	return c, c.Update(args)
-}
-
-// Run implements component.Component.
-func (c *Component) Run(ctx context.Context) error {
-	var cancel context.CancelFunc
-	for {
-		select {
-		case <-ctx.Done():
-			return nil
-		case <-c.newDiscoverer:
-			// cancel any previously running discovery
-			if cancel != nil {
-				cancel()
-			}
-			// function to send updates on change
-			f := func(t []discovery.Target) {
-				c.opts.OnStateChange(Exports{Targets: t})
-			}
-			// create new context so we can cancel it if we get any future updates
-			// since it is derived from the main run context, it only needs to be
-			// canceled directly if we receive new updates
-			newCtx, cancelFunc := context.WithCancel(ctx)
-			cancel = cancelFunc
-
-			// finally run discovery
-			c.discMut.Lock()
-			disc := c.latestDisc
-			c.discMut.Unlock()
-			go discovery.RunDiscovery(newCtx, disc, f)
-		}
-	}
-}
-
-// Update implements component.Compnoent.
-func (c *Component) Update(args component.Arguments) error {
-	newArgs := args.(SDConfig)
-
-	disc, err := promk8s.New(c.opts.Logger, newArgs.Convert())
-	if err != nil {
-		return err
-	}
-	c.discMut.Lock()
-	c.latestDisc = disc
-	c.discMut.Unlock()
-
-	select {
-	case c.newDiscoverer <- struct{}{}:
-	default:
-	}
-
-	return nil
 }
